unify model metadata: auto-populate cards from provider APIs, extract real context/tokens
- Remove hardcoded 2-model ModelCardStore defaults, populate from draw-through cache
- Extract real context_length, max_completion_tokens, pricing from provider API responses
- Infer tags (vision, reasoning, coding, fast, tools, general) from model names
- Surface all tags as capabilities in /api/show for Copilot/clients
- Parse parameter size from model names (regex, no hardcoded list)
- Add cache.find() with suffix matching for flexible model lookup

+++ src/keymux/types.rs
@@ -17,7 +17,7 @@ pub struct ModelInfo {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct WebModelCard {
     pub tags: Vec<String>,
-    pub context_window: u32,
+    pub context_window: u64,
     pub pricing: Option<Pricing>,
     pub reasoning_depth: u8, // 1-10
     pub code_native: bool,
